requirements.txt
librosa==0.11.0
numpy==2.3.2
pynput==1.8.1
scikit_learn==1.7.1
tensorflow_cpu==2.18.1
